requirements.txt
Flask
tensorflow
numpy
pillow
datetime
cassandra-driver
matplotlib
